pangea/pyproject.toml
[tool.poetry]
name = "pangea"
version = "0.1.0"
description = ""
authors = ["Damian Krystkiewicz <damian.krystkiewicz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.2"
django-environ = "^0.4.5"
tqdm = "^4.45.0"
pexpect = "^4.8.0"
loguru = "^0.4.1"
fire = "^0.3.1"
flake8 = "^3.7.9"
jinja2 = "^2.11.2"
plasma-comm = "^0.1.6"
black = "^19.10b0"
envo = "^0.4.8"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
mypy = "^0.770"
pep8-naming = "^0.10.0"
pytest = "^5.4.1"

[tool.poetry.scripts]
pangea = "pangea.scripts:_main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
